% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Inspect_Data}
\alias{Inspect_Data}
\title{Data inspection and quality check}
\usage{
Inspect_Data(auto = T, prm = F)
}
\arguments{
\item{auto}{Logical. If set to TRUE, the function takes in default values for certain inputs. Check the
scope document for the list of defaults.}

\item{prm}{Logical. If set to TRUE, the functions filters the data for an additional category. For the US market,
use prm = TRUE to filter for both Retailer and Promo Family. For the India market, can be used to filter for both
Brand and Cluster in case consolidated data for multiple brands is provided in a single file.}
}
\description{
The function reads the input raw data, records summary statistics for the data and generates a report to
summarise data quality and characteristics.
}
