% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Prepare_Data}
\alias{Prepare_Data}
\title{Automated data processing and preparation for BBN model creation}
\usage{
Prepare_Data(prm = F)
}
\arguments{
\item{prm}{Logical. If set to TRUE, the functions filters the data for an additional category. For the US market,
use prm = TRUE to filter for both Retailer and Promo Family. For the India market, can be used to filter for both
Brand and Cluster in case consolidated data for multiple brands is provided in a single file.}
}
\value{
A list containing the aggregated processed data, the time period and scope document information along
with the path of the working directory. These are read as input in the Transform_Data function.
}
\description{
The function reads the input raw data, provides the user a set of options to treat the data and writes out an
aggregated data with all the variables treated.
}
