% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Transform_Data}
\alias{Transform_Data}
\title{Automated data transformation and variable selection}
\usage{
Transform_Data(ls_dt_sub, vec_time, dt_scp, wd_path, prm = F)
}
\arguments{
\item{ls_dt_sub}{The "Data" object returned by Prepare_Data}

\item{vec_time}{The "Period" object returned by Prepare_Data}

\item{dt_scp}{The "Scope" object returned by Prepare_Data}

\item{wd_path}{The "Working Directory" object returned by Prepare_Data}

\item{prm}{Logical. If set to TRUE, the functions filters the data for an additional category. For the US market,
use prm = TRUE to filter for both Retailer and Promo Family. For the India market, can be used to filter for both
Brand and Cluster in case consolidated data for multiple brands is provided in a single file.}
}
\value{
All outputs are generated within specific folders for the different Markets/Retailers and Promos
}
\description{
The function reads the processed data from Prepare_Data, creates transformed versions of selected variables,
selects the transformed version with the ideal correlation to the target and writes out the final train and
test splits of the data which can be directly fed to the Bayesian network creation module or any other models.
}
